test/test_api.py
# tests/backend/test_api.py (Example)
import pytest
from flask import Flask
# ... Other imports


@pytest.fixture  #Pytest fixture to create a test client for the Flask API, for making requests to the API and testing different routes and endpoints.
def client():
    app = Flask(__name__)  #Or import your actual app.
    # Configure app for testing (e.g. database, routes)


    with app.test_client() as client:  #Uses `app.test_client()` to create test client.
        yield client



def test_get_agents(client):  #Example test case for your get agents route.  Modify route if needed.
    response = client.get('/agents')  #Makes GET request to '/agents' endpoint.
    assert response.status_code == 200 #Check if the status code is correct
    data = json.loads(response.data) #Parse response
    assert isinstance(data, list)   #Checks return data type, ensures it's a list, or whatever the intended type is.  This verifies that the backend returns data in the expected format and structure.  If the returned data is not as expected, this will raise an assertion error, causing the test to fail, and allowing you to quickly discover and diagnose problems with your API routes or endpoints.


# ... (Other tests for your API routes - /tasks, /agents (POST), /agents/<id> (PUT, DELETE), /login, etc.)


# tests/backend/test_agent.py (Example)
import pytest
from agents.agent import Agent, WebSearcherAgent  # Import necessary components and classes, etc.
from unittest.mock import Mock # For mocking



#Fixture for initializing an LLM
@pytest.fixture

def mock_llm_interface():
  llm_interface = Mock(spec=LLM_Interface) #Create mock LLM


  llm_interface.generate_text.return_value = "Mock LLM response." #Sets return value to test responses.
  return llm_interface


@pytest.fixture
def agent(mock_llm_interface):
    return WebSearcherAgent(
        agent_id=0, name="Test Agent", description="A test agent", skills=["web_search"], 
        tools=["web_search"], model_path="mock_path", api_key="mock_key", search_engine_id="mock_id"  #Mock agent values

    ) #Instantiate agent



def test_web_search_agent_handle_task(agent, mock_llm_interface):
    task = "Search the web for information about the moon."

    response = agent.handle_task(task)

    mock_llm_interface.generate_text.assert_called_once()  # Assert LLM was called

    assert "Mock LLM response" in response # Check for part of the response

# ... more tests (DocumentExpert, etc.)


# ... (tests for other modules: tools, memory manager, etc.)